+++ app/javascript/application.js
@@ -3,6 +3,5 @@ import "controllers";
 import Rails from "@rails/ujs";
 import "imageEditor"; 
 import "pageReload";
-import "paintSearch";
 
 Rails.start();
