+++ config/importmap.rb
@@ -9,4 +9,3 @@ pin "@rails/ujs", to: "@rails--ujs.js" # @7.1.3
 pin "application"
 pin "imageEditor", to: "imageEditor.js", preload: true
 pin "pageReload", to: "pageReload.js", preload: true
-pin "paintSearch", to: "paintSearch.js", preload: true
